+++ kitty/state.h
@@ -465,6 +465,10 @@ typedef struct OSWindow {
     CloseRequest close_request;
     bool is_layer_shell, hide_on_focus_loss;
     struct { int x, y; } last_drag_event;
+    struct {
+        double pending_pixels_x, pending_pixels_y;
+        int last_v120_dir_x, last_v120_dir_y;
+    } scroll;
 } OSWindow;
 
 static inline float